+++ packages/x25519/src/Cipher/algorithms/recommended.ts
@@ -2,7 +2,5 @@
  * Copyright (c) 2019-2020 Digital Bazaar, Inc. All rights reserved.
  */
 import * as cipher from './xc20p';
-import * as keyAgreement from './x25519';
 
 export { cipher };
-export { keyAgreement };
